producer_kraken/requirements.txt
kafka-python>=2.2.0
websockets>=12.0
